side bar component created
