+++ frontend/src/App.tsx
@@ -1,21 +1,32 @@
+// App.tsx
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Login from "./components/Login";
-import Dashboard from "./components/Dashboard";
-import ProtectedRoute from "./components/ProtectedRoute";
+import Login from "./views/login/Login";
+import Dashboard from "./views/dashboard";
+import Clients from "./views/clients";
+import Layout from "./views/components/Layout";
+import ProtectedRoute from "./views/ProtectedRoute";
 import { ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
 
 function App() {
   return (
     <Router>
       <Routes>
+        {/* Public */}
         <Route path="/" element={<Login />} />
-        <Route path="/dashboard" element={
-          <ProtectedRoute>
-            <Dashboard />
+
+        {/* Protected with sidebar */}
+        <Route
+          element={
+            <ProtectedRoute>
+              <Layout />
             </ProtectedRoute>
-          } />
+          }
+        >
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/clients" element={<Clients />} />
+        </Route>
       </Routes>
+
       <ToastContainer position="top-right" autoClose={2000} />
     </Router>
   );
