+++ frontend/src/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import userReducer from "./userSlice";
+import userReducer from "./UserSlice";
 
 export const store = configureStore({
   reducer: {
